Adds a basic main menu and HUD
Adds widgets for a main menu and in-game HUD. All the artwork is just placeholder art, naturally.

Adds a MainMenu level, for the main menu obviously. Also modifies the TestArea level to add the HUD to the viewport.

Currently the HUD has no actual functionality other than lighting up when you click the buttons. The main menu allows you to start a new game, which takes you to TestArea, or exit. The settings in the options menu do nothing, they're just there for the sake of example.